# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Uni Ride Sharing API",
    "description": "API documentation for the university ride sharing application",
    "version": "1.0.0"
  },
  "host": "localhost:3002",
  "basePath": "/api",
  "schemes": ["http", "https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "sessionAuth": {
      "type": "apiKey",
      "name": "Cookie",
      "in": "header",
      "description": "Session cookie for authentication"
    }
  },
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "User login",
        "description": "Authenticates a user and creates a session",
        "tags": ["Authentication"],
        "parameters": [
          {
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["email", "password", "userType"],
              "properties": {
                "email": {
                  "type": "string",
                  "format": "email",
                  "example": "john@example.com"
                },
                "password": {
                  "type": "string",
                  "format": "password",
                  "example": "password123"
                },
                "userType": {
                  "type": "string",
                  "enum": ["student", "driver", "admin"],
                  "example": "student"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/rides": {
      "get": {
        "summary": "Get available rides",
        "description": "Gets a list of available rides based on search criteria",
        "tags": ["Rides"],
        "security": [{"sessionAuth": []}],
        "parameters": [
          {
            "name": "pickupLocation",
            "in": "query",
            "type": "string",
            "description": "Starting location"
          },
          {
            "name": "dropoffLocation",
            "in": "query",
            "type": "string",
            "description": "Destination location"
          },
          {
            "name": "departureTime",
            "in": "query",
            "type": "string",
            "format": "date-time",
            "description": "Departure time"
          },
          {
            "name": "maxPrice",
            "in": "query",
            "type": "number",
            "description": "Maximum price"
          }
        ],
        "responses": {
          "200": {
            "description": "List of matching rides"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      
      "post": {
        "summary": "Create a new ride",
        "description": "Creates a new ride offer (drivers only)",
        "tags": ["Rides"],
        "security": [{"sessionAuth": []}],
        "parameters": [
          {
            "name": "rideData",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["pickupLocation", "dropoffLocation", "departureTime", "fare", "seatsAvailable"],
              "properties": {
                "pickupLocation": {
                  "type": "string",
                  "example": "Roehampton University"
                },
                "dropoffLocation": {
                  "type": "string",
                  "example": "Putney"
                },
                "departureTime": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2023-04-15T14:30:00Z"
                },
                "fare": {
                  "type": "number",
                  "example": 15.50
                },
                "seatsAvailable": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 8,
                  "example": 4
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Ride created successfully"
          },
          "400": {
            "description": "Invalid input data"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not a driver"
          }
        }
      }
    },
    "/rides/{rideId}": {
      "get": {
        "summary": "Get ride details",
        "description": "Gets details for a specific ride",
        "tags": ["Rides"],
        "security": [{"sessionAuth": []}],
        "parameters": [
          {
            "name": "rideId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the ride"
          }
        ],
        "responses": {
          "200": {
            "description": "Ride details"
          },
          "404": {
            "description": "Ride not found"
          }
        }
      }
    },
...
    "/rewards": {
      "get": {
        "summary": "Get rewards",
        "description": "Gets available rewards for the current user",
        "tags": ["Rewards"],
        "security": [{"sessionAuth": []}],
        "responses": {
          "200": {
            "description": "List of available rewards and points"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/rewards/claim/{rewardId}": {
      "post": {
        "summary": "Claim a reward",
        "description": "Claims a reward with user points",
        "tags": ["Rewards"],
        "security": [{"sessionAuth": []}],
        "parameters": [
          {
            "name": "rewardId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID of the reward"
          }
        ],
        "responses": {
          "200": {
            "description": "Reward claimed successfully"
          },
          "400": {
            "description": "Not enough points"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Reward not found"
          }
        }
      }
    }
  }
} 